Trivial cleanups to get rid of GCC warnings.
src/fontset.c (Fset_fontset_font): Removed unused variables.
src/fringe.c (Ffringe_bitmaps_at_pos): Ditto.
src/xdisp.c (Flookup_image_map, note_mode_line_or_margin_highlight): Ditto.
src/xterm.c (x_detect_focus_change, x_calc_absolute_position): Ditto.

src/intervals.h (compare_string_intervals): New declaration.

src/intervals.h (NULL_INTERVAL_P): Shut up GCC warnings by casting the
left side of the comma expression to void.
src/lisp.h (CHECK): Ditto.

src/keyboard.c (modify_event_symbol): Shut up GCC warning about format
template type mismatch.


git-archimport-id: lorentey@elte.hu--2004/emacs--multi-tty--0--patch-120

+++ src/keyboard.c
@@ -6248,10 +6248,10 @@ modify_event_symbol (symbol_num, modifiers, symbol_kind, name_alist_or_stem,
 	  char *buf = (char *) alloca (len + 50);
 	  if (sizeof (int) == sizeof (EMACS_INT))
 	    sprintf (buf, "%s-%d", SDATA (name_alist_or_stem),
-		     XINT (symbol_int) + 1);
+		     (int)XINT (symbol_int) + 1);
 	  else if (sizeof (long) == sizeof (EMACS_INT))
 	    sprintf (buf, "%s-%ld", SDATA (name_alist_or_stem),
-		     XINT (symbol_int) + 1);
+		     (long)XINT (symbol_int) + 1);
 	  value = intern (buf);
 	}
       else if (name_table != 0 && name_table[symbol_num])
